requirements.txt
Django==1.5.1
Fabric==1.5.3
Pillow==2.0.0
South==0.8
argparse==1.2.1
boto==2.8.0
django-annoying==0.7.7
django-autoslug==1.6.1
django-nose==1.1
django-storages==1.1.6
djangorestframework==2.2.0
eventlet==0.12.1
greenlet==0.4.0
gunicorn==0.17.4
nose==1.3.0
paramiko==1.9.0
progressbar==2.3
psycopg2==2.4.6
pycrypto==2.6
raven==3.1.13
requests==1.1.0
wsgiref==0.1.2
yolk==0.4.3
sorl-thumbnail==11.12
-e git+https://github.com/toastdriven/django-haystack.git@c5e0ce5221fc97f6a9a6fd9d6b6fad6aec960842#egg=django_haystack-dev
pyelasticsearch==0.3
django-oauth2-provider==0.2.4
twilio==3.4.5
gevent==0.13.8
django-ses==0.4.1
hoover==0.6.0
django-debug-toolbar==0.9.4
pytz==2013b
django-kronos==0.3